+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# changelog
+# Changelog
 
+* 2021-02-16 Add subsliceFrom(), subsliceInRrange(), splitWithLimit(), stringByReplacingStringInRange() in StringSlice & NcString.
 * 2021-02-15 Add NcObject::isKindOf()